update Code Generator

+++ tools/pydia2code/pydia2code.py
@@ -3,12 +3,11 @@ import libxslt
 import yaml
 import os
 import glob
-
-# TODO Header template
-# TODO Header decorator
+import datetime
 
 print("=== PyDia2Code C++ Header-only generator === ")
-print("By Timothe PEREZ | 3IS 2015 @ ENSEA | www.ensea.fr")
+print("Version 1.0.0")
+print("By Timothe Perez | 3IS 2015 @ ENSEA | www.ensea.fr")
 
 print("= Step 0 : Initialization")
 file = file("config.yml", "r")
@@ -22,16 +21,33 @@ print("Generating Headers for " + config["project"]["name"] +
 tmp_folder = config["directories"]["temp"]
 inp_folder = config["directories"]["input"]
 out_folder = config["directories"]["output"]
+dia2xml = config["stylesheets"]["dia2xml"]
+xml2cpp = config["stylesheets"]["xml2cpp"]
 #
+
+print("Checking folders existence")
+if not os.path.exists(tmp_folder):
+    print("TMP folder not found. Creating")
+    os.mkdir(tmp_folder)
+if not os.path.exists(out_folder):
+    print("Out folder not found. Creating")
+    os.mkdir(out_folder)
+for namespace in config["namespaces"]:
+    if not os.path.exists(out_folder + namespace + "/"):
+        print("Out folder for " + namespace + " not found. Creating")
+        os.mkdir(out_folder + namespace + "/")
+if not os.path.exists(inp_folder):
+    raise IOError("Input folder not found")
+print("Done")
 
 print("Cleaning folders...")
 for namespace in config["namespaces"]:
-    for f in glob.glob(out_folder + namespace + "/*.h"):
+    for f in glob.glob(out_folder + namespace + "/*.hpp"):
         os.unlink(f)
 print("Done")
-# input("Wait")
+
 print("= Step 1 : extracting data from DIA file (Non Gzipped!)")
-styledoc = libxml2.parseFile("dia-uml.xsl")
+styledoc = libxml2.parseFile(dia2xml)
 style = libxslt.parseStylesheetDoc(styledoc)
 for namespace in config["namespaces"]:
     try:
@@ -48,22 +64,25 @@ doc.freeDoc()
 result.freeDoc()
 
 print("= Step 2 : creating headers files from XML")
-styledoc = libxml2.parseFile("dia-uml2cpp.xsl")
+styledoc = libxml2.parseFile(xml2cpp)
 style = libxslt.parseStylesheetDoc(styledoc)
-
+params = {"project": repr(config["project"]["name"]),
+          "author": repr(config["author"]),
+          "version": repr(config["project"]["version"]),
+          "date": repr(datetime.datetime.now().strftime("%Y-%m-%d %H:%M"))}
 for namespace in config["namespaces"]:
     try:
         print("Headers for namespace " + namespace)
-        params = {'directory': repr(out_folder + namespace + "/")}
+        params["directory"] = repr(out_folder + namespace + "/")
+        params["include"] = repr(config["includes"][namespace])
 
         doc = libxml2.parseFile(tmp_folder + namespace + ".xml")
         result = style.applyStylesheet(doc, params)
-        # style.saveResultToFilename("engine", result, 0)
         print(namespace + " : XML2CPP OK")
     except:
         print(namespace + " : Fail at XML2CPP")
 
-print("Warning, XLST lib can fail even if no exception is thrown.")
-print("Make sure your headers are correctly generated")
+print("Warning generation can fail even if no exception is thrown.")
+print("Make sure your headers are correctly generated.")
 
 exit(0)
